Bots/Reddit/requirements.txt
praw>=7.7
python-dotenv>=1.0
pyyaml>=6.0